refactor(pool-page): extract _tabs/reserves-tab
Move ReservesTab (~165 lines) out of page.tsx into _tabs/reserves-tab.tsx.

Behavior-preserving — verified via:
- pnpm test (1366/1366 green)
- pnpm typecheck (clean)
- chrome-devtools snapshot of /pool/<id>?tab=reserves vs baseline:
  zero structural changes, console clean

page.tsx: 2169 → 2008 lines (-161).
